add hibernate.cfg.xml and dependency to POM-file; remove Tender.java to model.response dir
