+++ src/main/java/project/model/dao/HTenderDAO.java
@@ -1,0 +1,32 @@
+package project.model.dao;
+
+import java.util.List;
+import java.util.Optional;
+
+public class HTenderDAO implements ITenderDAO {
+    @Override
+    public Optional get(long id) {
+
+        return Optional.empty();
+    }
+
+    @Override
+    public List getAll() {
+        return null;
+    }
+
+    @Override
+    public void save(Object o) {
+
+    }
+
+    @Override
+    public void update(Object o, String[] params) {
+
+    }
+
+    @Override
+    public void remove(Object o) {
+
+    }
+}
