+++ src/main/java/project/model/dto/etender/EtenderDTO.java
@@ -1,0 +1,5 @@
+package project.model.dto.etender;
+
+//TODO Create EtenderDTO classes
+public class EtenderDTO {
+}
