+++ src/main/java/project/model/response/Initiator.java
@@ -1,0 +1,43 @@
+package project.model.response;
+
+public class Initiator {
+    private int id;
+    private String name;
+    private String contactName;
+    private String contactPhone;
+
+    public Initiator() {
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getContactName() {
+        return contactName;
+    }
+
+    public void setContactName(String contactName) {
+        this.contactName = contactName;
+    }
+
+    public String getContactPhone() {
+        return contactPhone;
+    }
+
+    public void setContactPhone(String contactPhone) {
+        this.contactPhone = contactPhone;
+    }
+}
