+++ src/main/java/project/model/response/Site.java
@@ -1,0 +1,36 @@
+package project.model.response;
+
+public class Site {
+    private int id;
+    private String name;
+    private String url;
+
+    public Site() {
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getUrl() {
+        return url;
+    }
+
+    public void setUrl(String url) {
+        this.url = url;
+    }
+}
+
+
